feat: can now draw strokes on the page, still bugged tho

+++ src/diary/models/point.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 @dataclass
 class Point:
-    def __init__(self, x: int, y: int, pressure: float):
-        self.x: int = x
-        self.y: int = y
+    def __init__(self, x: float, y: float, pressure: float):
+        self.x: float = x
+        self.y: float = y
         self.pressure: float = pressure
